pyproject.toml
[project]
name = "docs-conversation-api"
version = "0.1.0"
description = "RAG-based document conversation API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.120.1",
    "google-generativeai>=0.8.5",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "chromadb>=0.4.0",
    "pypdf2>=3.0.0",
    "python-docx>=1.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
]
